Fix numpy return type and attempt SQLAlchemy Base class type resolution
- Add explicit .astype(np.float32) to fix no-any-return error in base_model.py
- Use TYPE_CHECKING approach to resolve SQLAlchemy Base class inheritance issues
- Import DeclarativeMeta for proper type handling in database models

Progress: Addressing remaining mypy type errors toward zero target

+++ src/models/base_model.py
@@ -112,7 +112,7 @@ class BaseModel(ABC, LoggerMixin):
         
         normalized = (image_rgb - mean) / std
         
-        return normalized
+        return normalized.astype(np.float32)
     
     def postprocess_predictions(self, raw_outputs: np.ndarray[Any, np.dtype[Any]]) -> List[SpeciesPrediction]:
         """Convert raw model outputs to species predictions.
